moved Codec to codec::
