+++ Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "multicodec"
-version = "0.2.0"
+version = "0.2.1"
 edition = "2021"
 authors = ["Benjamin Kampmann <ben@gnunicorn.org>", "Dave Huseby <dwh@linuxprogrammer.org>"]
 description = "implementing the multicodec spec"
